no key deletion on dummy

+++ Class/wireguard.py
@@ -100,9 +100,10 @@ class Wireguard(Base):
     def setInterface(self,file,state):
         self.cmd(f'bash {self.path}/links/{file}.sh {state}')
 
-    def cleanInterface(self,interface):
+    def cleanInterface(self,interface,deleteKey=True):
         os.remove(f"{self.path}/links/{interface}.sh")
-        os.remove(f"{self.path}/links/{interface}.key")
+        if deleteKey:
+            os.remove(f"{self.path}/links/{interface}.key")
 
     def saveFile(self,data,path):
         with open(path, 'w') as file: file.write(data)
@@ -162,6 +163,10 @@ class Wireguard(Base):
                     print(f"Got {req.status_code} with {req.text} aborting")
             except Exception as ex:
                 exit(ex)
-        #clean dummy
-        self.setInterface("dummy","down")
-        self.cleanInterface("dummy")
+        #load configs
+        configs = self.getConfigs()
+        #check for dummy
+        if not self.hasDummy(configs):
+            #clean dummy
+            self.setInterface("dummy","down")
+            self.cleanInterface("dummy",False)
